% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benmR.R
\docType{package}
\name{benmR}
\alias{benmR}
\alias{benmR-package}
\title{benmR: A package for estimating ecological niches through a Bayesian model.}
\description{
The underlying Bayesian model that is used to estimate ecological niches with
this package is presented in Jimenez \emph{et al} (2018?). The main
\code{benmR} package is to provide code to apply this methodology of niche
purpose of the estimation. It package provides three categories of functions
grouped together according to their main purpose: visualization, modeling and
estimation, and analysis of results.
}
\section{Visualization functions}{

The package contains a function called \code{plotdata} which displays the
data to be used to estimate the ecological niche of a species.
Once the Bayesian estimation is done for the parameters of interest,
\code{PlotIterations} plots the fitted ellipsoids that come from the
simulations of the posterior distribution on top of the background data in
environmental space.
The function \code{UncertaintyPlot} calculates an uncertainty index for each
cell in the environmental space were the niche is being estimated and
displays these values as a color map.
}

\section{Modeling functions}{

The modeling functions are those specifically designed to define the
components of the posterior distribution. \code{priorpar} fixes the parameters
of the a priori distributions using tolerance ranges of the species.
\code{Supp}, \code{Initth}, and \code{Energy} are the functions needed to
simulate from the posterior distribution by using the t-walk algorithm,
being the last function the one that contains the expression for the posterior.
The function \code{postsample} is the one used to simulate from the posterior
and depends on all the other functions mentioned in this section.
}

\section{Analysis functions}{

These are functions that help processing the results of the t-walk sampling.
Under development...
}

