% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-tools.R
\name{plotdata}
\alias{plotdata}
\title{Visualization of datasets to be used for niche estimatition.}
\usage{
plotdata(back, occ.sp, tolran, sp.col)
}
\arguments{
\item{back}{numeric matrix that contains the background data for the species.}

\item{occ.sp}{numeric matrix that contains the species occurrence data.}

\item{tolran}{numeric matrix with one row and four columns named: xleft,
ybottom, xright, ytop.}

\item{sp.col}{character strign with a color name identifying the species.}
}
\value{
A plot with two panels. The panel on the left will show the data in
  geographical space and the panel on the right will show the data in
  environmental space.
}
\description{
\code{plotdata} returns two plots that display the data needed to estimate the
niche of a species in both geographical and environmental space.
}
\details{
This function produce a visualization of the background points, the
occurrence data and the tolerance ranges of the species of interest, in both
the geographical and the environmental spaces (displaying only the firs two
or three dimensions of the latter).
The first two columns of \code{back} must contain the longitude and latitude
coordinates of all the points in the grid that define the region of interest,
the remaining columns must contain the values (at least two) of the
environmental variables at each location. Similarly, the first two columns of
\code{occu.sp} must contain the geographical coordinates for the occurrence
data, and the remaining columns must contain the environmental combinations.
\code{tolran} must provide the tolerance limits (lower and upper bounds) for
each environmental variable odd columns contain lower limits even columns
containd the upper limits.
}
\examples{
plotdata(back1,occsp1,tolran1,"tomato")
}
