% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior-parmtr-defntn.R
\name{priorpar}
\alias{priorpar}
\title{Defining the a priori parameters for the model.}
\usage{
priorpar(tolran, nsd, alpha = 2)
}
\arguments{
\item{tolran}{numeric vector containing the tolerance limits of the species.}

\item{nsd}{number of standard deviations covered by a tolerance range.}

\item{alpha}{number of degrees of freedom in the a priori Wishart distribution.}
}
\value{
A list containing a vector of length equals to the number of
  dimensions in the environmental space, and two squared matrices with that
  same number of rows/columns. The first two elements represent the a priori
  parameters of \code{mu} and the third element is the scale matrix that
  defines the a priori distribution of \code{A}.
}
\description{
\code{priorpar} uses the physiological tolerance limits of the species to
determine the values of the parameters that define the a priori distribution.
}
\details{
\code{tolran} must contain tolerance ranges of the species to every
environmental variable that defines the environmental space. The lenght of
this vector must be twice the number of environmental variables. The values
contained in odd positions represent lower tolerance limits, and the values
in even positions are the upper limits.
The value of \code{nsd} is used to estimate the diagonal entries of the
covariance matrix that defines the a priori (multivariate normal)
distribution of the parameter A. It is set such that each range is
approximately equals to \code{nsd} times the corresponding standard
deviation.
The value of \code{alpha} should be changed only when working with an
environmental space of more than 2 dimensions.
}
\examples{
# 2-dimensional case with range covering 10 standard deviations
tolran1 <- c(1,2,-1,1)
priorpar(tolran1,10)
}
